PR checks / Validate: minor improvements
This switches to combining the optional checks into one step each for
better clarity.

It also uses `git add .` to ensure that change that just adds an
untracked file will be noticed. That shouldn’t be an issue for either
optional check, but it might make changes clearer for somebody looking
at the failure report.